Remove actors where they have no state.

+++ src/main/scala/model/Price.scala
@@ -6,9 +6,9 @@ import java.time.LocalDate
   * Created by dennis on 18/8/16.
   */
 case class Price(date: LocalDate,
-                 open: BigDecimal,
-                 high: BigDecimal,
-                 low: BigDecimal,
-                 close: BigDecimal,
+                 open: Double,
+                 high: Double,
+                 low: Double,
+                 close: Double,
                  volume: Long,
-                 adjusted: BigDecimal)
+                 adjusted: Double)
